안드로이드에서 댓글 작성시 게시글 번호 데이터 받아오기 성공(DB 저장 미구현)

+++ ccp_(0329_오전1032취합1)/ccp/app/src/main/java/com/example/ccp/model/CommentDTO.kt
@@ -1,11 +1,9 @@
 package com.example.ccp.model
 
-import java.time.LocalDateTime
-
 data class CommentDTO(
     var cnum: Long? = null,
     var writerUsername: String? = null,
     var content: String? = null,
     var regdate: String? = null,
-    var boardBnum: Long? = null
+    var boardBnum: Int? = null
 )
